feat: move rxdb of planner to new db
