+++ src/app/[lang]/(mods-pages)/student/planner/components/bulk-actions/bulk-actions-menu.tsx
@@ -15,7 +15,7 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { SemesterDocType } from "@/config/rxdb";
+import { SemesterDocType } from "@/app/[lang]/(mods-pages)/student/planner/rxdb";
 import { CourseStatus } from "../../types";
 
 interface BulkActionsMenuProps {
