+++ src/app/[lang]/(mods-pages)/student/planner/data/semesters.ts
@@ -1,4 +1,4 @@
-import { SemesterDocType } from "@/config/rxdb";
+import { SemesterDocType } from "@/app/[lang]/(mods-pages)/student/planner/rxdb";
 import { CourseStatus } from "../types";
 import { RxCollection } from "rxdb";
