+++ src/app/[lang]/(mods-pages)/student/planner/planner-settings.tsx
@@ -30,7 +30,7 @@ import {
   ItemDocType,
   PlannerDataDocType,
   SemesterDocType,
-} from "@/config/rxdb";
+} from "./rxdb";
 import { useRxCollection } from "rxdb-hooks";
 import {
   createPlannerData,
