+++ src/app/[lang]/(mods-pages)/student/planner/types.ts
@@ -1,3 +1,3 @@
-import { FolderDocType } from "@/config/rxdb";
+import { FolderDocType } from "./rxdb";
 
 export type CourseStatus = "completed" | "in-progress" | "planned" | "failed";
